responsible-ai-moderationlayer/src/requirements/requirement.txt
pyyaml
regex
python-dotenv
pymongo
sqlalchemy
openai
hvac
azure-keyvault-secrets
azure-identity
# nltk==3.9
langchain-community
langchain-openai
setuptools==70.0.0
flask
flask[async]
flask-cors
azure-ai-translation-text==1.0.0b1
waitress
hypercorn
flask-swagger-ui
langcodes
Pillow
demoji
pytest
grapheme
wheel==0.38.1
aiohttp
nltk==3.9.1
boto3
langchain-aws
google-generativeai
fkscore
../lib/better_profanity-2.0.0-py3-none-any.whl
../lib/en_core_web_lg-3.7.1-py3-none-any.whl
